requirements.txt
streamlit
streamlit_option_menu
azure-storage-blob
azure-cosmos
openai
azure-identity
pillow
python-dotenv
